fix: moved wrapRootElement to each file
