+++ packages/gatsby-theme-newrelic/gatsby-browser.js
@@ -1,2 +1,15 @@
+export const wrapRootElement = ({ element }, pluginOptions) => {
+  return (
+    <LayoutContext.Provider value={pluginOptions.layout}>
+      {pluginOptions.splitio ? (
+        <SplitIOProvider config={getSplitConfig(pluginOptions)}>
+          {element}
+        </SplitIOProvider>
+      ) : (
+        element
+      )}
+    </LayoutContext.Provider>
+  );
+};
+
 export { default as wrapPageElement } from './gatsby/wrap-page-element';
-export { default as wrapRootElement } from './gatsby/wrap-root-element';
