+++ packages/gatsby-theme-newrelic/gatsby-ssr.js
@@ -1,5 +1,12 @@
 import wrapPageElement from './gatsby/wrap-page-element';
 import onPreRenderHTML from './gatsby/on-pre-render-html';
-import wrapRootElement from './gatsby/wrap-root-element';
+
+const wrapRootElement = ({ element }, pluginOptions) => {
+  return (
+    <LayoutContext.Provider value={pluginOptions.layout}>
+      {element}
+    </LayoutContext.Provider>
+  );
+};
 
 export { onPreRenderHTML, wrapPageElement, wrapRootElement };
